requirements.txt
scikit-learn>=1.2.0
python-docx>=1.0.0